requirements.txt
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
cattrs==1.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
facebook-sdk==3.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flask-paginate==0.8.1
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.17
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-facebook-api==0.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
responses==0.12.1
scipy==1.6.1
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
tinysegmenter==0.3
tldextract==3.1.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
types-requests==2.25.0
types-six==0.1.7
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
